pyproject.toml
[tool.poetry]
name = "luckypot"
version = "0.1.0"
description = "Another pygame engine. It might be what you need, if you're lucky to be called Diego."
authors = ["ddorn <diego.dorn@free.fr>"]
license = "GPL-3.0"
packages = [
    { include = "luckypot" },
]

include = [
    "luckypot/examples/assets/*",
]


[tool.poetry.dependencies]
python = "^3.9"
pygame-ce = "^2.5.0"
numpy = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
